feat(processing_engine): respond to PR feedback, put python code behind system-py feature flag
