+++ influxdb3_catalog/Cargo.toml
@@ -15,7 +15,6 @@ iox_time.workspace = true
 # Local deps
 influxdb3_id = { path = "../influxdb3_id" }
 influxdb3_wal = { path = "../influxdb3_wal" }
-influxdb3_processing_engine = { path = "../influxdb3_processing_engine" }
 
 # crates.io dependencies
 arrow.workspace = true
