--- influxdb3_processing_engine/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "influxdb3_processing_engine"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-serde = { version = "1.0.214", features = ["derive"] }
-serde_json = "1.0.132"
-influxdb3_wal = { path = "../influxdb3_wal" }
-
-[lints]
-workspace = true
