--- influxdb3_processing_engine/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod processing_engine_plugins;
