+++ influxdb3_py_api/Cargo.toml
@@ -5,16 +5,20 @@ authors.workspace = true
 edition.workspace = true
 license.workspace = true
 
+
+[features]
+system-py = ["pyo3"]
 [dependencies]
 influxdb3_wal = { path = "../influxdb3_wal" }
 influxdb3_catalog = {path = "../influxdb3_catalog"}
-influxdb-line-protocol = { workspace = true }
 schema = { workspace = true }
 
 [dependencies.pyo3]
 version = "0.23.3"
 # this is necessary to automatically initialize the Python interpreter
 features = ["auto-initialize"]
+optional = true
+
 
 [lints]
 workspace = true
