+++ influxdb3_server/Cargo.toml
@@ -39,7 +39,6 @@ influxdb3_write = { path = "../influxdb3_write" }
 iox_query_influxql_rewrite = { path = "../iox_query_influxql_rewrite" }
 influxdb3_sys_events = { path = "../influxdb3_sys_events" }
 influxdb3_telemetry = { path = "../influxdb3_telemetry" }
-influxdb3_processing_engine = { path = "../influxdb3_processing_engine" }
 
 # crates.io Dependencies
 anyhow.workspace = true
