+++ influxdb3_write/Cargo.toml
@@ -5,6 +5,9 @@ authors.workspace = true
 edition.workspace = true
 license.workspace = true
 
+[features]
+"system-py" = ["influxdb3_py_api/system-py"]
+
 [dependencies]
 # Core Crates
 data_types.workspace = true
@@ -23,7 +26,6 @@ influxdb3_cache = { path = "../influxdb3_cache" }
 influxdb3_catalog = { path = "../influxdb3_catalog" }
 influxdb3_id = { path = "../influxdb3_id" }
 influxdb3_test_helpers = { path = "../influxdb3_test_helpers" }
-influxdb3_processing_engine = { path = "../influxdb3_processing_engine"}
 influxdb3_wal = { path = "../influxdb3_wal" }
 influxdb3_telemetry = { path = "../influxdb3_telemetry" }
 influxdb3_py_api = {path = "../influxdb3_py_api"}
