+++ influxdb3_write/src/write_buffer/queryable_buffer.rs
@@ -14,14 +14,14 @@ use datafusion::catalog::Session;
 use datafusion::common::DataFusionError;
 use datafusion::logical_expr::Expr;
 use datafusion_util::stream_from_batches;
-use hashbrown::{HashMap, HashSet};
+use hashbrown::HashMap;
 use influxdb3_cache::last_cache::LastCacheProvider;
 use influxdb3_cache::meta_cache::MetaCacheProvider;
 use influxdb3_cache::parquet_cache::{CacheRequest, ParquetCacheOracle};
 use influxdb3_catalog::catalog::{Catalog, DatabaseSchema};
 use influxdb3_id::{DbId, TableId};
-use influxdb3_processing_engine::processing_engine_plugins::TriggerSpecification;
-use influxdb3_py_api::PyWriteBatch;
+#[cfg(feature = "system-py")]
+use influxdb3_py_api::system_py::PyWriteBatch;
 use influxdb3_wal::{CatalogOp, SnapshotDetails, WalContents, WalFileNotifier, WalOp, WriteBatch};
 use iox_query::chunk_statistics::{create_chunk_statistics, NoColumnRanges};
 use iox_query::exec::Executor;
@@ -38,7 +38,6 @@ use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::oneshot;
 use tokio::sync::oneshot::Receiver;
-use TriggerSpecification::SingleTableWalWrite;
 
 #[derive(Debug)]
 pub struct QueryableBuffer {
@@ -509,55 +508,60 @@ impl BufferState {
             .db_schema_by_id(&write_batch.database_id)
             .expect("database should exist");
 
-        let write_tables: HashSet<_> = write_batch
-            .table_chunks
-            .keys()
-            .map(|key| {
-                let table_name = db_schema.table_map.get_by_left(key).unwrap();
-                table_name.to_string()
-            })
-            .collect();
-
-        let triggers: Vec<_> = db_schema
-            .processing_engine_triggers
-            .values()
-            .filter_map(|trigger| match &trigger.trigger {
-                SingleTableWalWrite { table_name } => {
-                    if write_tables.contains(table_name.as_str()) {
-                        Some((trigger, vec![table_name.clone()]))
-                    } else {
-                        None
+        // TODO: factor this out
+        #[cfg(feature = "system-py")]
+        {
+            use influxdb3_wal::TriggerSpecificationDefinition;
+            use influxdb3_wal::TriggerSpecificationDefinition::SingleTableWalWrite;
+            let write_tables: hashbrown::HashSet<_> = write_batch
+                .table_chunks
+                .keys()
+                .map(|key| {
+                    let table_name = db_schema.table_map.get_by_left(key).unwrap();
+                    table_name.to_string()
+                })
+                .collect();
+            let triggers: Vec<_> = db_schema
+                .processing_engine_triggers
+                .values()
+                .filter_map(|trigger| match &trigger.trigger {
+                    SingleTableWalWrite { table_name } => {
+                        if write_tables.contains(table_name.as_str()) {
+                            Some((trigger, vec![table_name.clone()]))
+                        } else {
+                            None
+                        }
                     }
-                }
-                TriggerSpecification::AllTablesWalWrite => {
-                    if !write_tables.is_empty() {
-                        Some((
-                            trigger,
-                            write_tables.iter().map(ToString::to_string).collect(),
-                        ))
-                    } else {
-                        None
+                    TriggerSpecificationDefinition::AllTablesWalWrite => {
+                        if !write_tables.is_empty() {
+                            Some((
+                                trigger,
+                                write_tables.iter().map(ToString::to_string).collect(),
+                            ))
+                        } else {
+                            None
+                        }
                     }
-                }
-            })
-            .collect();
-        if !triggers.is_empty() {
-            // Create PyWriteBatch instance
-            let py_write_batch = PyWriteBatch {
-                write_batch: write_batch.clone(),
-                schema: db_schema.clone(),
-            };
-            for (trigger, write_tables) in triggers {
-                for table in &write_tables {
-                    if let Err(err) = py_write_batch.call_against_table(
-                        table,
-                        trigger.plugin.code.as_str(),
-                        trigger.plugin.function_name.as_str(),
-                    ) {
-                        error!(
-                            "failed to call trigger {} with error {}",
-                            trigger.trigger_name, err
-                        )
+                })
+                .collect();
+            if !triggers.is_empty() {
+                // Create PyWriteBatch instance
+                let py_write_batch = PyWriteBatch {
+                    write_batch: write_batch.clone(),
+                    schema: db_schema.clone(),
+                };
+                for (trigger, write_tables) in triggers {
+                    for table in &write_tables {
+                        if let Err(err) = py_write_batch.call_against_table(
+                            table,
+                            trigger.plugin.code.as_str(),
+                            trigger.plugin.function_name.as_str(),
+                        ) {
+                            error!(
+                                "failed to call trigger {} with error {}",
+                                trigger.trigger_name, err
+                            )
+                        }
                     }
                 }
             }
